Add files via upload
This portal lets anyone with a used book to share its details including pictures. Any user accessing the system can either search for a book or books available in his vicinity. User can further narrow his search based on category and/or author of the book. Contact details of the owner are provided to registered users. Once a book is marked as sold, it will not appear in further searches